+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add support for skipping and un-skipping tests using the test item context menu
 * Automatically recompile code in debug/normal mode when starting/stopping debugging
 * Add support for the debug toolbar with additional buttons for common port commands
+* Add partial support for the "Variables" and "Call Stack" panes in the "Run and Debug" sidebar
 * Add "Help: Logtalk Handbook" command to open the Logtalk Handbook using Live Preview if installed, or the default browser otherwise
 * Add multi-root workspace support for code navigation, refactoring, testing, documentation, and chat participant features
 * Add "Use implicit message sending" refactoring for replacing explicit message sending calls with a `uses/2` directive